refactor(indexer): split graphql and postgres store modules
